refactor(projection)!: convert projectIntoSink into rxjs operator
RxJS already provides a good way for composing observables,
abstracting it away just adds unnecessary coupling.

Also extract `logProjectionProgress` into a standalone operator.
