+++ packages/projection/src/util.ts
@@ -1,0 +1,4 @@
+import { PointOrOrigin } from '@cardano-sdk/core';
+
+export const pointDescription = (point: PointOrOrigin) =>
+  point === 'origin' ? 'origin' : `slot ${point.slot}, block ${point.hash}`;
